moved low-level API to dedicated package
