+++ mini_al/mini_al.go
@@ -1,5 +1,5 @@
 // Package mal - Mini audio library (mini_al cgo bindings).
-package mal
+package mini_al
 
 /*
 #cgo CFLAGS: -std=gnu99
